introduce react-hook-form for better state management
